requirements.txt
asgiref==3.10.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
psutil==5.9.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
